api/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
psycopg2-binary
pyjwt
bcrypt
python-dotenv
